+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 # 2.19.5 (unreleased)
 
+* API:
+    * Introduce version flexibility in workflow-import endpoint (\#3154).
 * Database:
     * Enforce non-duplication constraints for `TaskGroupV2` and `TaskGroupActivityV2` via `unique` database checks (\#3185).